preparation for multi device type support
