+++ .gitignore
@@ -122,3 +122,6 @@ PyViCare/__init__.py
 *.local.sh
 env*
 dump.*json
+
+# PyCharm
+.idea
