--- PyViCare/PyViCareHybrid.py
@@ -1,6 +1,0 @@
-from PyViCare.PyViCareGazBoiler import GazBoiler
-from PyViCare.PyViCareHeatPump import HeatPump
-
-
-class Hybrid(GazBoiler, HeatPump):
-    pass
