+++ PyViCare/generic/PyViCareGenericDevice.py
@@ -1,0 +1,10 @@
+from PyViCare.heating.PyViCareHeatingDevice import HeatingDevice
+
+
+# TODO: This needs to be updated if a new device type is added!
+class GenericDevice(HeatingDevice):
+    """
+    This class needs to extend all different types of devices in order
+    to work as a generic class.
+    """
+    pass
