+++ PyViCare/heating/PyViCareGazBoiler.py
@@ -1,11 +1,11 @@
 from typing import Any, List
 
-from PyViCare.PyViCareDevice import (Device, DeviceWithComponent,
-                                     get_available_burners)
+from PyViCare.heating.PyViCareHeatingDevice import (HeatingDevice, HeatingDeviceWithComponent,
+                                                    get_available_burners)
 from PyViCare.PyViCareUtils import handleNotSupported
 
 
-class GazBoiler(Device):
+class GazBoiler(HeatingDevice):
 
     @property
     def burners(self) -> List[Any]:
@@ -264,7 +264,7 @@ class GazBoiler(Device):
         return self.service.getProperty("heating.power.consumption.summary.dhw")["properties"]["lastYear"]["value"]
 
 
-class GazBurner(DeviceWithComponent):
+class GazBurner(HeatingDeviceWithComponent):
 
     @property
     def burner(self) -> str:
