+++ PyViCare/heating/PyViCareHeatCurveCalculation.py
