+++ PyViCare/heating/PyViCareHybrid.py
@@ -1,0 +1,6 @@
+from PyViCare.heating.PyViCareGazBoiler import GazBoiler
+from PyViCare.heating.PyViCareHeatPump import HeatPump
+
+
+class Hybrid(GazBoiler, HeatPump):
+    pass
