+++ PyViCare/sensor/PyViCareRoomSensor.py
