+++ tests/test_TestForMissingProperties.py
@@ -2,7 +2,7 @@ import json
 import re
 import unittest
 from os import listdir
-from os.path import dirname, isfile, join
+from os.path import dirname, isfile, isdir, join
 
 from tests.helper import readJson
 
@@ -116,17 +116,29 @@ class TestForMissingProperties(unittest.TestCase):
 
     def read_all_python_code(self):
         python_path = join(dirname(__file__), '../PyViCare')
-        python_files = [f for f in listdir(python_path) if isfile(join(python_path, f))]
+        # stores pairs of [path, filename]
+        python_files = []
+        # searches in all subdirectories
+        self.get_all_files(python_files, python_path)
 
         all_python_files = {}
 
         for python in python_files:
-            if not python.endswith(".py"):
+            if not python[1].endswith(".py"):
                 continue
 
-            with open(join(python_path, python)) as f:
-                all_python_files[python] = f.read()
+            with open(join(python[0], python[1])) as f:
+                all_python_files[python[1]] = f.read()
+
         return all_python_files
+
+    def get_all_files(self, files, path):
+        for f in listdir(path):
+            new_path = join(path, f)
+            if isdir(new_path):
+                self.get_all_files(files, new_path)
+            elif isfile(new_path):
+                files.append([path, f])
 
     def read_all_features(self):
         response_path = join(dirname(__file__), './response')
