+++ tests/test_Vitocal222S.py
@@ -1,6 +1,6 @@
 import unittest
 
-from PyViCare.PyViCareHeatPump import HeatPump
+from PyViCare.heating.PyViCareHeatPump import HeatPump
 from tests.helper import now_is
 from tests.ViCareServiceMock import ViCareServiceMock
