+++ tests/test_Vitocaldens222F.py
@@ -1,6 +1,6 @@
 import unittest
 
-from PyViCare.PyViCareHybrid import Hybrid
+from PyViCare.heating.PyViCareHybrid import Hybrid
 from tests.ViCareServiceMock import ViCareServiceMock
