+++ tests/test_Vitodens222W.py
@@ -1,6 +1,6 @@
 import unittest
 
-from PyViCare.PyViCareGazBoiler import GazBoiler
+from PyViCare.heating.PyViCareGazBoiler import GazBoiler
 from PyViCare.PyViCareUtils import PyViCareNotSupportedFeatureError
 from tests.ViCareServiceMock import ViCareServiceMock
